[refactor] Move MyApp & delete unused code.
