+++ test/widget_test.dart
@@ -6,6 +6,7 @@
 // tree, read text, and verify that the values of widget properties are correct.
 
 import 'package:flutter/material.dart';
+import 'package:flutter_pixabay/app.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:flutter_pixabay/main.dart';
